Refactored merkleComputer initialization
